requirements.txt
# PersonaSteer V2 - 依赖包列表
# 用于 AI 人物引导/角色控制研究项目

# 核心深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
Pillow>=9.5.0
opencv-python>=4.8.0

# 机器学习工具
scikit-learn>=1.3.0

# 模型管理与训练
transformers>=4.36.0
timm>=0.9.0
deepspeed>=0.12.0
accelerate>=0.25.0
datasets>=2.16.0

# 实验跟踪
tensorboard>=2.15.0
wandb>=0.15.0

# 可视化
matplotlib>=3.8.0
seaborn>=0.13.0

# 配置管理
pyyaml>=6.0
omegaconf>=2.3.0

# 日志和进度
tqdm>=4.65.0
loguru>=0.7.0

# 文件操作和路径
pathlib2>=2.3.0
python-dotenv>=1.0.0

# 类型检查和代码质量
typing-extensions>=4.5.0

# LLM API
openai>=1.0.0

# 开发和调试
ipython>=8.12.0
pytest>=7.4.0
ruff>=0.1.0
